try half precision

+++ visda_classification/basenet.py
@@ -109,21 +109,21 @@ class AlexNet2(nn.Module):
     def __init__(self):
         super(AlexNet2, self).__init__()
         self.features = nn.Sequential(
-            nn.Conv2d(3, 64, kernel_size=5, stride=4, padding=2),#nn.Conv2d(3, 64, kernel_size=11, stride=4, padding=2),
-            nn.ReLU(inplace=True),
-            nn.MaxPool2d(kernel_size=3, stride=2),
-            nn.Conv2d(64, 192, kernel_size=5, padding=2),
-            nn.ReLU(inplace=True),
-            nn.MaxPool2d(kernel_size=3, stride=2),
-            nn.Conv2d(192, 384, kernel_size=3, padding=1),
-            nn.ReLU(inplace=True),
-            nn.Conv2d(384, 256, kernel_size=3, padding=1),
-            nn.ReLU(inplace=True),
-            nn.Conv2d(256, 256, kernel_size=3, padding=1),
-            nn.ReLU(inplace=True),
-            nn.MaxPool2d(kernel_size=3, stride=2),
+            nn.Conv2d(3, 64, kernel_size=5, stride=4, padding=2).half(),#nn.Conv2d(3, 64, kernel_size=11, stride=4, padding=2),
+            nn.ReLU(inplace=True).half(),
+            nn.MaxPool2d(kernel_size=3, stride=2).half(),
+            nn.Conv2d(64, 192, kernel_size=5, padding=2).half(),
+            nn.ReLU(inplace=True).half(),
+            nn.MaxPool2d(kernel_size=3, stride=2).half(),
+            nn.Conv2d(192, 384, kernel_size=3, padding=1).half(),
+            nn.ReLU(inplace=True).half(),
+            nn.Conv2d(384, 256, kernel_size=3, padding=1).half(),
+            nn.ReLU(inplace=True).half(),
+            nn.Conv2d(256, 256, kernel_size=3, padding=1).half(),
+            nn.ReLU(inplace=True).half(),
+            nn.MaxPool2d(kernel_size=3, stride=2).half(),
         )
-        self.avgpool = nn.AdaptiveAvgPool2d((6, 6))
+        self.avgpool = nn.AdaptiveAvgPool2d((6, 6)).half()
     def forward(self, x):
         x = self.features(x)
         x = self.avgpool(x)
